authentication is working

+++ client/src/pages/ListPage.js
@@ -1,9 +1,9 @@
 import "./ListPage.css"
 import { useStoreContext } from "../utils/GlobalState";
-import { REMOVE_LIST_ITEM, ADD_FAVORITE, UPDATE_LISTS, LOADING, SET_CURRENT_LIST, ADD_LIST } from "../utils/actions"
-import React, { useState, useEffect, useRef } from "react";
-import { Link, useParams } from "react-router-dom";
-import { Form, Button, Table, ThemeProvider } from 'react-bootstrap';
+import { REMOVE_LIST_ITEM, UPDATE_LISTS, LOADING, SET_CURRENT_LIST, ADD_LIST } from "../utils/actions"
+import React, { useEffect } from "react";
+import { useParams } from "react-router-dom";
+import { Button, Table, ThemeProvider } from 'react-bootstrap';
 import API from "../utils/API";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCheck } from '@fortawesome/free-solid-svg-icons';
@@ -12,10 +12,11 @@ import CreateListForm from "../components/CreateListForm"
 import { faHome } from '@fortawesome/free-solid-svg-icons';
 
 // import codename from ZOEYTHING
-export function ListPage() {
+export function ListPage(props) {
     const { codename } = useParams();
     const { id } = useParams();
 
+    console.log("this is props ==>", props)
     // const [formObject, setFormObject] = useState([]);
 
     const [state, dispatch] = useStoreContext();
@@ -37,32 +38,44 @@ export function ListPage() {
     const addFavorite = (e) => {
         e.preventDefault();
         console.log("THIS IS STATE.CURRENTLIST FROM fave BUTTONE", state)
-        API.addFavorite(codename, {
-            favorite: true,
-        }).then(result => {
-            console.log("This is the API result from addFavorite", result)
-            dispatch({
-                type: UPDATE_LISTS,
-                list: state.currentList
-            });
-        })
-
+        const faveList = state.currentList.favorites;
+        if (faveList.includes(props.user.email)) {
+            return null
+        } else {
+            faveList.push(props.user.email)
+            console.log(faveList)
+            API.addFavorite(codename, {
+                favorites: faveList,
+            }).then(result => {
+                console.log("This is the API result from addFavorite", result)
+                dispatch({
+                    type: UPDATE_LISTS,
+                    list: state.currentList
+                });
+            })
+        }
     };
 
     const deleteFavorite = (e) => {
         e.preventDefault();
-        API.deleteFavorite(codename, {
-            favorite: false,
-        }).then(result => {
-            console.log("This is the API result from deleteFavorite", result)
-            dispatch({
-                type: UPDATE_LISTS,
-                list: state.currentList
-            });
-        })
-
+        const faveList = state.currentList.favorites;
+        if (faveList.includes(props.user.email)) {
+            const index = faveList.indexOf(props.user.email)
+            faveList.splice(index, 1, "")
+            console.log(faveList)
+            API.deleteFavorite(codename, {
+                favorites: faveList,
+            }).then(result => {
+                console.log("This is the API result from deleteFavorite", result)
+                dispatch({
+                    type: UPDATE_LISTS,
+                    list: state.currentList
+                });
+            })
+        } else {
+            return null
+        }
     }
-
 
     const getList = (codename) => {
         dispatch({ type: LOADING });
@@ -112,6 +125,7 @@ export function ListPage() {
                         {/* <h1>Your codename is:</h1>
                         <h2>{codename}</h2>
                         <br></br> */}
+                        <h5 id="user">You are signed in as {props.user.email}</h5>
                         <CreateListForm />
                         {/* <div className="faveButtonContainer">
                             <button className="saveFaveButton" onClick={addFavorite}>Save to favorites</button>
@@ -124,74 +138,74 @@ export function ListPage() {
                         {console.log("This is state -->", state)}
                         <ThemeProvider prefixes={{ table: 'my-table' }}>
                             <div className="tableContainer">
-                            <Table 
-                            bordered 
-                            responsive 
-                            size="sm">
-                                <thead>
-                                    <tr>
-                                        <th>#</th>
-                                        <th>Item Name
+                                <Table
+                                    bordered
+                                    responsive
+                                    size="sm">
+                                    <thead>
+                                        <tr>
+                                            <th>#</th>
+                                            <th>Item Name
                                     </th>
-                                        <th>Quantity</th>
-                                        <th>Purchased?</th>
-                                        {/* <th>ID</th> */}
-                                    </tr>
-                                    {console.log(state.currentList)}
-                                </thead>
-
-                                {/* This checks to see if there are any items, if not -> does nothing, if it does -> return table body */}
-                                {state.currentList.items ? <tbody>
-                                    {console.log("this is state.currentList.items ----> ", state.currentList.items)}
-                                    {console.log("this is state ----> ", state)}
-                                    {state.currentList.items.map(item =>
-                                        <tr>
-                                            <td className="countCell">{count++}</td>
-                                            <td>{item.itemName}</td>
-                                            <td>{item.quantity}</td>
-                                            <td><Button style={{ display: "flex", justifyContent: "center", margin: "auto", marginTop: ".15rem", marginBottom: ".15rem", height: "1.5rem", width: "1.5rem", alignItems: "center" }} onClick={() => {
-                                                removeListItem(item._id)
-                                            }}>
-                                                <FontAwesomeIcon style={{ height: ".5rem", width: ".5rem" }} icon={faCheck}>
-                                                    {item.purchased}</FontAwesomeIcon>
-                                            </Button>
-                                            </td>
-                                            {/* <td>{item._id}</td> */}
+                                            <th>Quantity</th>
+                                            <th>Purchased?</th>
+                                            {/* <th>ID</th> */}
                                         </tr>
-                                    )}
+                                        {console.log(state.currentList)}
+                                    </thead>
 
-                                </tbody> : null
+                                    {/* This checks to see if there are any items, if not -> does nothing, if it does -> return table body */}
+                                    {state.currentList.items ? <tbody>
+                                        {console.log("this is state.currentList.items ----> ", state.currentList.items)}
+                                        {console.log("this is state ----> ", state)}
+                                        {state.currentList.items.map(item =>
+                                            <tr>
+                                                <td className="countCell">{count++}</td>
+                                                <td>{item.itemName}</td>
+                                                <td>{item.quantity}</td>
+                                                <td><Button style={{ display: "flex", justifyContent: "center", margin: "auto", marginTop: ".15rem", marginBottom: ".15rem", height: "1.5rem", width: "1.5rem", alignItems: "center" }} onClick={() => {
+                                                    removeListItem(item._id)
+                                                }}>
+                                                    <FontAwesomeIcon style={{ height: ".5rem", width: ".5rem" }} icon={faCheck}>
+                                                        {item.purchased}</FontAwesomeIcon>
+                                                </Button>
+                                                </td>
+                                                {/* <td>{item._id}</td> */}
+                                            </tr>
+                                        )}
 
-
-                                }
-
-                                {state.currentList.length > 0 ? <tbody>
-                                    {console.log("this is state.currentList ----> ", state.currentList)}
-                                    {console.log("this is state ----> ", state)}
-                                    {state.currentList.map(item =>
-                                        <tr>
-                                            <td>{count++}</td>
-                                            <td>{item.itemName}</td>
-                                            <td>{item.quantity}</td>
-                                            <td><Button className="purchasedButton" onClick={() => {
-                                                removeListItem(item._id)
-                                            }}>
-                                                <FontAwesomeIcon icon={faCheck}>
-                                                    {item.purchased}</FontAwesomeIcon>
-                                            </Button>
-                                            </td>
-                                            {/* <td>{item._id}</td> */}
-                                        </tr>
-                                    )}
-
-                                </tbody> : null
+                                    </tbody> : null
 
 
-                                }
+                                    }
+
+                                    {state.currentList.length > 0 ? <tbody>
+                                        {console.log("this is state.currentList ----> ", state.currentList)}
+                                        {console.log("this is state ----> ", state)}
+                                        {state.currentList.map(item =>
+                                            <tr>
+                                                <td>{count++}</td>
+                                                <td>{item.itemName}</td>
+                                                <td>{item.quantity}</td>
+                                                <td><Button className="purchasedButton" onClick={() => {
+                                                    removeListItem(item._id)
+                                                }}>
+                                                    <FontAwesomeIcon icon={faCheck}>
+                                                        {item.purchased}</FontAwesomeIcon>
+                                                </Button>
+                                                </td>
+                                                {/* <td>{item._id}</td> */}
+                                            </tr>
+                                        )}
+
+                                    </tbody> : null
+
+
+                                    }
 
 
 
-                            </Table>
+                                </Table>
                             </div>
                             <br></br>
                             <div className="faveButtonContainer">
